:recycle: Move OperandTypeSelection into its own component
